requirements.txt
pandas==2.1.4
numpy==1.25.0
scikit-learn==1.3.2
flask==3.0.0
flask-cors==4.0.0
plotly==5.17.0
dash==2.16.1
dash-bootstrap-components==1.5.0
google-generativeai==0.3.2
python-dotenv==1.0.0
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
imbalanced-learn==0.11.0
google-generativeai==0.3.2